+++ mayan/apps/checkouts/widgets.py
@@ -1,4 +1,3 @@
-from collections import OrderedDict
 import datetime
 
 from django import forms
@@ -14,19 +13,17 @@ class SplitTimeDeltaWidget(NamedMultiWidget):
     A Widget that splits a timedelta input into two field: one for unit of
     time and another for the amount of units.
     """
-    def __init__(self, attrs=None):
-        widgets = OrderedDict()
-        widgets['unit'] = forms.widgets.Select(
+    subwidgets = {
+        'unit': forms.widgets.Select(
             attrs={'style': 'width: 8em;'}, choices=TIME_DELTA_UNIT_CHOICES
-        )
-        widgets['amount'] = forms.widgets.NumberInput(
+        ),
+        'amount': forms.widgets.NumberInput(
             attrs={
                 'maxlength': 4, 'style': 'width: 8em;',
                 'placeholder': _('Amount')
             }
         )
-
-        super(SplitTimeDeltaWidget, self).__init__(widgets=widgets, attrs=attrs)
+    }
 
     def decompress(self, value):
         return {